+++ packages/implementations/react/src/composite/SideNavigation/src/index.ts
@@ -1,0 +1,5 @@
+import { SideNavigation } from "./SideNavigation";
+
+export { SideNavigation } from "./SideNavigation";
+export type { SideNavigationProps } from "./SideNavigation";
+export default SideNavigation;