+++ packages/implementations/react/src/composite/Sidebar/src/index.ts
@@ -1,0 +1,5 @@
+import { Sidebar } from "./Sidebar";
+
+export { Sidebar } from "./Sidebar";
+export type { SidebarProps } from "./sidebar.types";
+export default Sidebar;