+++ packages/implementations/react/src/composite/Sidebar/src/sidebar.spec.tsx
@@ -1,0 +1,33 @@
+import React from "react";
+import { render, screen } from "@testing-library/react";
+
+import { Sidebar } from "./Sidebar";
+import { SidebarProps } from "./sidebar.types";
+
+const makeSut = (rest: SidebarProps) => {
+  render(<Sidebar {...rest} data-testid="sidebar-element" />);
+};
+
+describe("GIVEN <Sidebar />", () => {
+  describe("WHEN rendered", () => {
+    it("THEN should correctly render the submitted content", () => {
+      makeSut({ children: <div>My content</div>, open: true });
+      expect(screen.getByText("My content")).toBeDefined();
+    });
+  });
+
+  describe("THEN should correctly render or not based on the open property", () => {
+    it("THEN should correctly render if open is true", () => {
+      makeSut({ children: <div>My content</div>, open: true });
+      expect(screen.getByTestId("overlay-sidebar-button")).toBeTruthy();
+      expect(
+        screen.getByTestId("sidebar-element").getAttribute("class"),
+      ).toContain("open_true");
+    });
+
+    it("THEN should not render if open is false", () => {
+      makeSut({ children: <div>My content</div>, open: false });
+      expect(screen.queryByTestId("overlay-sidebar-button")).toBeNull();
+    });
+  });
+});