+++ packages/implementations/react/src/atomic/Tooltip/CHANGELOG.md
@@ -9,8 +9,8 @@ Tooltip are informative, specific, and action -oriented text labels that provide
 - Added `enabledHover`, `content`, `children`, `visible`, `onVisibility`, `arrow`, `matchReferenceWidth`, `position`, `enabledClick`, `enabledDismiss`, `offset`, `padding` and `inverted` properties to the component. ([#32](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/32) by [@junior](https://git.rarolabs.com.br/junior))
 - Added stories on Component. ([#32](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/32) by [@junior](https://git.rarolabs.com.br/junior))
 - Added `closeable`, `title`, `subtitle`, `children`, `onClose`, `inverted` and `padding` properties to the Component `Tooltip.Header`. ([#3](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/3)2 by [@junior](https://git.rarolabs.com.br/junior))
-- Added `children` and `padding` properties to the Component `Tooltip.Body`. ([#3](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/3)2 by [@junior](https://git.rarolabs.com.br/junior))
-- Added `children` and `padding` properties to the Component `Tooltip.Footer`. ([#3](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/3)2 by [@junior](https://git.rarolabs.com.br/junior))
+- Added `children` and `padding` properties to the Component `Tooltip.Body`. ([#32](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/32)2 by [@junior](https://git.rarolabs.com.br/junior))
+- Added `children` and `padding` properties to the Component `Tooltip.Footer`. ([#32](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/32)2 by [@junior](https://git.rarolabs.com.br/junior))
 
 <!-- #### 🛠 Breaking changes -->