+++ packages/styles/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 RarUI Styles deprive all styles needed to build components.
 
+## 2024-05-15 `1.18.0`
+
+#### 🎉 New features
+
+- Added new `backgroundPositionX` and `backgroundPositionY` properties to the component `Box`. ([#38](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/38) by [@junior](https://git.rarolabs.com.br/junior))
+- Added new style pack for `sideNavigation` component. ([#38](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/38) by [@junior](https://git.rarolabs.com.br/junior))
+- Added new style pack for `sidebar` component. ([#38](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/38) by [@junior](https://git.rarolabs.com.br/junior))
+
 ## 2024-05-07 `1.16.0`
 
 #### 🎉 New features